Rename style.py to styles.py
