+++ almanac/style/__init__.py
@@ -1,1 +1,1 @@
-from .style import DARK_MODE_STYLE, LIGHT_MODE_STYLE  # noqa
+from .styles import DARK_MODE_STYLE, LIGHT_MODE_STYLE  # noqa
